Finish 82 lesson. Load new films when scroll, work with database without internet

+++ app/src/main/java/com/example/myapplication/MainActivity.java
@@ -18,6 +18,7 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.CompoundButton;
+import android.widget.ProgressBar;
 import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -42,9 +43,14 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
     private TextView tvPopularity;
     private TextView tvTopRated;
     private MainViewModel viewModel;
+    private ProgressBar progressBarLoading;
 
     private static final int LOADER_ID = 123;
     private LoaderManager loaderManager;
+
+    private static int page = 1;
+    private static boolean isLoading = false;
+    private static int methodOfSort;
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -79,6 +85,7 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
         tvPopularity = findViewById(R.id.textViewPopularity);
         tvTopRated = findViewById(R.id.textViewTopRated);
         recyclerViewPosters = findViewById(R.id.recyclerViewPosters);
+        progressBarLoading = findViewById(R.id.progressBarLoading);
         recyclerViewPosters.setLayoutManager(new GridLayoutManager(this, 2));
         movieAdapter = new MovieAdapter();
         recyclerViewPosters.setAdapter(movieAdapter);
@@ -86,6 +93,7 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
         switchSort.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
+                page = 1;
                 setMethodOfSort(b);
             }
         });
@@ -104,7 +112,9 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
         movieAdapter.setOnReachEndListener(new MovieAdapter.onReachEndListener() {
             @Override
             public void onReachEnd() {
-                Toast.makeText(MainActivity.this, "End Reached", Toast.LENGTH_SHORT).show();
+                if (!isLoading) {
+                    downloadData(methodOfSort, page);
+                }
             }
         });
 
@@ -112,7 +122,9 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
         moviesFromLiveData.observe(this, new Observer<List<Movie>>() {
             @Override
             public void onChanged(List<Movie> movies) {
-                movieAdapter.setMovies(movies);
+                if (page == 1){
+                    movieAdapter.setMovies(movies);
+                }
             }
         });
     }
@@ -129,7 +141,6 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
     }
 
     public void setMethodOfSort(boolean isChecked) {
-        int methodOfSort;
         if (isChecked) {
             methodOfSort = NetworkUtils.POPULARITY;
             tvTopRated.setTextColor(getResources().getColor(R.color.colorAccent));
@@ -139,7 +150,7 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
             tvTopRated.setTextColor(getResources().getColor(R.color.white_color));
             tvPopularity.setTextColor(getResources().getColor(R.color.colorAccent));
         }
-        downloadData(methodOfSort, 1);
+        downloadData(methodOfSort, page);
     }
 
     private void downloadData(int methodOfSort, int page) {
@@ -153,6 +164,13 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
     @Override
     public Loader<JSONObject> onCreateLoader(int id, @Nullable Bundle args) {
         NetworkUtils.JSONLoader jsonLoader = new NetworkUtils.JSONLoader(this, args);
+        jsonLoader.setOnStartLoadingListener(new NetworkUtils.JSONLoader.OnStartLoadingListener() {
+            @Override
+            public void onStartLoading() {
+                progressBarLoading.setVisibility(View.VISIBLE);
+                isLoading = true;
+            }
+        });
         return jsonLoader;
     }
 
@@ -160,11 +178,18 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
     public void onLoadFinished(@NonNull Loader<JSONObject> loader, JSONObject data) {
         ArrayList<Movie> movies = JSONUtils.getMoviesFromJSON(data);
         if (movies != null && !movies.isEmpty()){
-            viewModel.deleteAllMovies();
+            if (page == 1) {
+                viewModel.deleteAllMovies();
+                movieAdapter.clear();
+            }
             for (Movie movie : movies){
                 viewModel.insertMovie(movie);
             }
+            movieAdapter.addMovies(movies);
+            page++;
         }
+        isLoading = false;
+        progressBarLoading.setVisibility(View.INVISIBLE);
         loaderManager.destroyLoader(LOADER_ID);
     }
